feat: more policy into own module
