+++ libs/@local/hashql/eval/src/postgres/mod.rs
@@ -30,8 +30,8 @@
 use core::{alloc::Allocator, fmt::Display};
 
 use hash_graph_postgres_store::store::postgres::query::{
-    self, Column, Expression, Identifier, SelectExpression, SelectStatement, Transpile as _,
-    WhereExpression, table::EntityTemporalMetadata,
+    self, Column, Expression, Identifier, SelectExpression, SelectStatement, WhereExpression,
+    table::EntityTemporalMetadata,
 };
 use hashql_core::{
     debug_panic,
@@ -42,7 +42,6 @@ use hashql_core::{
 use hashql_mir::{
     body::{
         Body,
-        basic_block::BasicBlockId,
         local::Local,
         terminator::{GraphRead, GraphReadBody, GraphReadHead, TerminatorKind},
     },
@@ -63,6 +62,7 @@ use self::{
 pub use self::{
     continuation::ContinuationField,
     parameters::{Parameter, ParameterIndex, ParameterValue, Parameters, TemporalAxis},
+    prepared::{PatchPreparedQuery, PreparedQueries, PreparedQuery, PreparedQueryPatch},
 };
 use crate::context::CodeGenerationContext;
 
@@ -71,6 +71,7 @@ mod continuation;
 pub(crate) mod error;
 mod filter;
 mod parameters;
+mod prepared;
 mod projections;
 mod traverse;
 mod types;
@@ -186,51 +187,6 @@ impl Display for ColumnDescriptor {
                 )
             }
         }
-    }
-}
-
-/// A fully-compiled SQL query ready for execution.
-///
-/// Contains the typed query AST ([`SelectStatement`]), the parameter catalog ([`Parameters`])
-/// for binding runtime values, and a column manifest ([`ColumnDescriptor`]s) that tells the
-/// bridge how to decode each result column.
-pub struct PreparedQuery<'heap, A: Allocator> {
-    pub vertex_type: VertexType,
-    pub parameters: Parameters<'heap, A>,
-    pub(crate) projections: Projections,
-    pub statement: SelectStatement,
-    pub columns: Vec<ColumnDescriptor, A>,
-}
-
-impl<A: Allocator> PreparedQuery<'_, A> {
-    pub fn transpile(&self) -> impl Display {
-        core::fmt::from_fn(|fmt| self.statement.transpile(fmt))
-    }
-}
-
-/// Registry of compiled SQL queries, indexed by definition and basic block.
-///
-/// The SQL lowering pass produces one [`PreparedQuery`] per [`GraphRead`]
-/// terminator in the MIR. This struct stores them contiguously in `queries`
-/// with `offsets` providing per-definition starting positions, so
-/// [`find`](Self::find) can locate the correct query for a given `(DefId,
-/// BasicBlockId)` pair.
-///
-/// [`GraphRead`]: hashql_mir::body::terminator::GraphRead
-pub struct PreparedQueries<'heap, A: Allocator> {
-    offsets: Box<DefIdSlice<usize>, A>,
-    queries: Vec<(BasicBlockId, PreparedQuery<'heap, A>), A>,
-}
-
-impl<'heap, A: Allocator> PreparedQueries<'heap, A> {
-    pub fn find(&self, body: DefId, block: BasicBlockId) -> Option<&PreparedQuery<'heap, A>> {
-        let start = self.offsets[body];
-        let end = self.offsets[body.plus(1)];
-
-        self.queries[start..end]
-            .iter()
-            .find(|(id, _)| *id == block)
-            .map(|(_, query)| query)
     }
 }
 
@@ -382,7 +338,10 @@ impl<'eval, 'ctx, 'heap, A: Allocator, S: BumpAllocator>
         }
     }
 
-    fn compile_graph_read_entity(&mut self, read: &GraphRead<'heap>) -> PreparedQuery<'heap, A>
+    fn compile_graph_read_entity(
+        &mut self,
+        read: &GraphRead<'heap>,
+    ) -> prepared::PreparedQuery<'heap, A>
     where
         A: Clone,
     {
@@ -473,11 +432,12 @@ impl<'eval, 'ctx, 'heap, A: Allocator, S: BumpAllocator>
             .where_expression(db.where_expression)
             .build();
 
-        PreparedQuery {
+        prepared::PreparedQuery {
             vertex_type: VertexType::Entity,
             parameters: db.parameters,
             statement: query,
             projections: db.projections,
+            auxiliary_parameters: Vec::new_in(self.alloc.clone()),
             columns,
         }
     }
@@ -485,7 +445,10 @@ impl<'eval, 'ctx, 'heap, A: Allocator, S: BumpAllocator>
     /// Compiles a [`GraphRead`] into a [`PreparedQuery`].
     ///
     /// [`GraphRead`]: hashql_mir::body::terminator::GraphRead
-    pub fn compile_graph_read(&mut self, read: &'ctx GraphRead<'heap>) -> PreparedQuery<'heap, A>
+    pub fn compile_graph_read(
+        &mut self,
+        read: &'ctx GraphRead<'heap>,
+    ) -> prepared::PreparedQuery<'heap, A>
     where
         A: Clone,
     {
@@ -495,7 +458,7 @@ impl<'eval, 'ctx, 'heap, A: Allocator, S: BumpAllocator>
     }
 
     #[expect(unsafe_code)]
-    pub fn compile(&mut self) -> PreparedQueries<'heap, A>
+    pub fn compile(&mut self) -> prepared::PreparedQueries<'heap, A>
     where
         A: Clone,
     {
@@ -526,6 +489,6 @@ impl<'eval, 'ctx, 'heap, A: Allocator, S: BumpAllocator>
             offsets[body_id.plus(1)] = queries.len();
         }
 
-        PreparedQueries { offsets, queries }
+        prepared::PreparedQueries { offsets, queries }
     }
 }
